+++ carla_ros_bridge/src/carla_ros_bridge/bridge.py
@@ -20,7 +20,27 @@ from carla_ros_bridge.map import Map
 import carla_ros_bridge.object_sensor as ObjectSensor
 
 
-class CarlaRosBridge(Parent):
+# these imports have to be at the end to resolve cyclic dependency
+from carla_ros_bridge.actor import Actor   
+from carla_ros_bridge.spectator import Spectator  
+from carla_ros_bridge.sensor import Sensor       
+from carla_ros_bridge.traffic import Traffic, TrafficLight   
+from carla_ros_bridge.vehicle import Vehicle    
+from carla_ros_bridge.lidar import Lidar    
+from carla_ros_bridge.gnss import Gnss    
+from carla_ros_bridge.ego_vehicle import EgoVehicle  
+from carla_ros_bridge.collision_sensor import CollisionSensor  
+from carla_ros_bridge.lane_invasion_sensor import LaneInvasionSensor  
+from carla_ros_bridge.camera import RgbCamera, DepthCamera, SemanticSegmentationCamera
+
+
+class TestActor():
+    def __init__(self, id, name, parent):
+        self.id = name
+        self.name = name
+        self.parent = parent
+        
+class CarlaRosBridge(object):
 
     """
     Carla Ros bridge
@@ -34,9 +54,9 @@ class CarlaRosBridge(Parent):
         :type carla_world: carla.World
         :param binding: binding
         """
+        self.actors = {}
+        self.carla_world = carla_world
         self.params = binding.get_parameters()
-        super(CarlaRosBridge, self).__init__(
-            carla_id=0, carla_world=carla_world, frame_id='/map')
 
         self.binding = binding
         self.timestamp_last_run = 0.0
@@ -50,7 +70,7 @@ class CarlaRosBridge(Parent):
         self.update_lock = threading.Lock()
         self.carla_world.on_tick(self._carla_time_tick)
 
-        self.map = Map(carla_world=self.carla_world, parent=self, topic='/map')
+        self.map = Map(carla_world=self.carla_world, topic='/map')
 
     def destroy(self):
         """
@@ -64,9 +84,9 @@ class CarlaRosBridge(Parent):
         """
         self.update_child_actors_lock.acquire()
         self.update_lock.acquire()
+        self.get_binding().signal_shutdown("")
         self.actor_list = []
         self.get_binding().loginfo("Exiting Bridge")
-        super(CarlaRosBridge, self).destroy()
 
     def get_param(self, key, default=None):
         """
@@ -80,17 +100,6 @@ class CarlaRosBridge(Parent):
         :rtype: string
         """
         return self.params.get(key, default)
-
-    def topic_name(self):
-        """
-        Function (override) to get the topic name of this root entity.
-
-        The topic root '/carla' is returned by this bridge class.
-
-        :return: the final topic name of this
-        :rtype: string
-        """
-        return "/carla"
 
     def _carla_time_tick(self, carla_timestamp):
         """
@@ -131,12 +140,85 @@ class CarlaRosBridge(Parent):
         """
         if not self.get_binding().is_shutdown():
             if self.update_child_actors_lock.acquire(False):
-                # cache actor_list once during this update-loop
-                self.actor_list = self.carla_world.get_actors()
-                self.update_child_actors()
+                self.update_actors()
                 # actors are only created/deleted around once per second
                 time.sleep(1)
                 self.update_child_actors_lock.release()
+
+    def update_actors(self):
+        carla_actors = self.carla_world.get_actors()
+        #Add new actors
+        for actor in carla_actors:
+            if actor.id not in self.actors.keys():
+                self.create_actor(actor)
+
+        #create list of carla actors ids
+        carla_actor_ids = []
+        for actor in carla_actors:
+            carla_actor_ids.append(actor.id)
+
+        #remove non-existing actors
+        for actor_id in self.actors.keys():
+            id_to_delete = None
+            if actor_id not in carla_actor_ids:
+                id_to_delete = actor_id
+
+            if id_to_delete:
+                self.get_binding().logwarn("Remove actor {}".format(id_to_delete))
+                del self.actors[id_to_delete]
+
+    def create_actor(self, carla_actor):
+        parent = None
+        if carla_actor.parent:
+            if carla_actor.parent.id in self.actors:
+                parent = self.actors[carla_actor.parent.id]
+            else:
+                parent = self.create_actor(carla_actor.parent)
+        
+        actor = None
+        if carla_actor.type_id.startswith('traffic'):
+            if carla_actor.type_id == "traffic.traffic_light":
+                actor = TrafficLight(carla_actor, parent, self.binding)
+            else:
+                actor = Traffic(carla_actor, parent, self.binding)
+        elif carla_actor.type_id.startswith("vehicle"):
+            if carla_actor.attributes.get('role_name') in self.get_param('ego_vehicle').get('role_name'):
+                actor = EgoVehicle(carla_actor, parent, self.binding)
+            else:
+                actor = Vehicle(carla_actor, parent, self.binding)
+        elif carla_actor.type_id.startswith("sensor"):
+            if carla_actor.type_id.startswith("sensor.camera"):
+                if carla_actor.type_id.startswith("sensor.camera.rgb"):
+                    actor = RgbCamera(carla_actor, parent, self.binding)
+                elif carla_actor.type_id.startswith("sensor.camera.depth"):
+                    actor = DepthCamera(carla_actor, parent, self.binding)
+                elif carla_actor.type_id.startswith("sensor.camera.semantic_segmentation"):
+                    actor = SemanticSegmentationCamera(carla_actor, parent, self.binding)
+                else:
+                    actor = Camera(carla_actor, parent, self.binding)
+            if carla_actor.type_id.startswith("sensor.lidar"):
+                actor = Lidar(carla_actor, parent, self.binding)
+            if carla_actor.type_id.startswith("sensor.other.gnss"):
+                actor = Gnss(carla_actor, parent, self.binding)
+            if carla_actor.type_id.startswith("sensor.other.collision"):
+                actor = CollisionSensor(carla_actor, parent, self.binding)
+            if carla_actor.type_id.startswith("sensor.other.lane_invasion"):
+                actor = LaneInvasionSensor(carla_actor, parent, self.binding)
+            else:
+                actor = Sensor(carla_actor, parent, self.binding)
+        elif carla_actor.type_id.startswith("spectator"):
+            actor = Spectator(carla_actor, parent, self.binding)
+        else:
+            actor = Actor(carla_actor, parent, self.binding)
+        
+        self.get_binding().logwarn("Created Actor-{}(id={}, parent_id={},"
+                            " type={}, prefix={}, attributes={}".format(
+                                actor.__class__.__name__, actor.get_id(),
+                                actor.get_parent_id(), carla_actor.type_id,
+                                actor.get_topic_prefix(), carla_actor.attributes))
+        with self.update_lock:
+            self.actors[carla_actor.id] = actor
+        return actor
 
     def run(self):
         """
@@ -181,3 +263,26 @@ class CarlaRosBridge(Parent):
         """
         """
         return self.binding
+
+    def update(self):
+        """
+        Virtual (non abstract) function to update this object.
+
+        Override this function if the derived class has to perform
+        some additional update tasks. But don't forget to forward the update
+        call to the super class, ensuring that this concrete function is called.
+
+        The update part of the parent class consists
+        of updating the children of this by:
+
+        update the exising children
+
+        :return:
+        """
+        for actor_id in self.actors.keys():
+            try:
+                self.actors[actor_id].update()
+            except RuntimeError as e:
+                self.get_binding().logwarn("Update actor {}({}) failed: {}".format(
+                    actor.__class__.__name__, actor_id, e))
+                continue